+++ include/fi_util.h
@@ -60,7 +60,7 @@
 #include <fi_signal.h>
 #include <fi_enosys.h>
 #include <fi_osd.h>
-#include <fi_indexer.h>
+#include <ofi_indexer.h>
 
 #include "rbtree.h"
