+++ libfabric.vcxproj
@@ -573,7 +573,7 @@
     <ClInclude Include="include\fi_enosys.h" />
     <ClInclude Include="include\ofi_file.h" />
     <ClInclude Include="include\ofi_iov.h" />
-    <ClInclude Include="include\fi_indexer.h" />
+    <ClInclude Include="include\ofi_indexer.h" />
     <ClInclude Include="include\fi_list.h" />
     <ClInclude Include="include\fi_lock.h" />
     <ClInclude Include="include\fi_mem.h" />
