+++ prov/sockets/include/sock.h
@@ -54,7 +54,7 @@
 #include <ofi_atomic.h>
 #include <ofi_mr.h>
 #include <fi_enosys.h>
-#include <fi_indexer.h>
+#include <ofi_indexer.h>
 #include <fi_rbuf.h>
 #include <fi_list.h>
 #include <ofi_file.h>
